Support eventually consistent endpoint metadata in Watchman
