+++ requirements.in
@@ -1,3 +1,4 @@
+apscheduler~=3.6
 Click~=7.0
 cachetools~=3.1
 h5py~=2.8
