+++ requirements.txt
@@ -10,6 +10,7 @@ aiodns==2.0.0
 aiohttp==3.5.4
 alabaster==0.7.12         # via sphinx
 aniso8601==6.0.0          # via flask-restplus
+apscheduler==3.6.0
 asn1crypto==0.24.0        # via cryptography
 astor==0.7.1              # via tensorflow
 async-timeout==3.0.1      # via aiohttp
@@ -66,7 +67,7 @@ pyjwt==1.7.1              # via adal
 pyparsing==2.4.0          # via packaging
 pyrsistent==0.15.1        # via jsonschema
 python-dateutil==2.8.0
-pytz==2019.1              # via babel, flask-restplus, influxdb, pandas
+pytz==2019.1              # via apscheduler, babel, flask-restplus, influxdb, pandas, tzlocal
 pyyaml==5.1
 recommonmark==0.5.0
 requests-oauthlib==1.2.0  # via kubernetes
@@ -75,7 +76,7 @@ rsa==4.0                  # via google-auth
 scikit-learn==0.20.3
 scipy==1.2.1              # via keras, scikit-learn
 simplejson==3.16.0
-six==1.12.0               # via absl-py, cryptography, flask-restplus, google-auth, grpcio, h5py, influxdb, jsonschema, keras, keras-preprocessing, kubernetes, mock, packaging, pip-tools, protobuf, pyrsistent, python-dateutil, sphinx, tensorboard, tensorflow, tensorflow-estimator, websocket-client
+six==1.12.0               # via absl-py, apscheduler, cryptography, flask-restplus, google-auth, grpcio, h5py, influxdb, jsonschema, keras, keras-preprocessing, kubernetes, mock, packaging, pip-tools, protobuf, pyrsistent, python-dateutil, sphinx, tensorboard, tensorflow, tensorflow-estimator, websocket-client
 snowballstemmer==1.2.1    # via sphinx
 sphinx-click==2.1.0
 sphinx-rtd-theme==0.4.3
@@ -87,6 +88,7 @@ tensorflow==1.13.1
 termcolor==1.1.0          # via tensorflow
 typing-extensions==3.7.2  # via aiohttp
 typing==3.6.6             # via aiodns
+tzlocal==1.5.1            # via apscheduler
 urllib3==1.24.2
 websocket-client==0.56.0  # via kubernetes
 werkzeug==0.15.2          # via flask, tensorboard
